Make broker TCP destination option platform-neutral
